Fix minor issues revealed with -Weverythinig
Flags on apple clang 12.0.0:
```
-Wall -Weverything -Wno-c++98-compat -Wno-c++98-compat-pedantic
-Wno-missing-noreturn -Wno-global-constructors -Wno-padded
-Wno-documentation-unknown-command -pedantic -Wno-exit-time-destructors
-Wno-sign-conversion -Wno-weak-vtables -Wno-float-equal
-Wno-used-but-marked-unused -fdiagnostics-color=always
```

+++ src/random/cuda/RngStateStore.cc
@@ -21,7 +21,7 @@ namespace celeritas
 /*!
  * Construct with the number of RNG states.
  */
-RngStateStore::RngStateStore(size_type size, unsigned long host_seed)
+RngStateStore::RngStateStore(size_type size, unsigned int host_seed)
     : data_(size)
 {
     CELER_EXPECT(celeritas::device());
